% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metabomapr.R
\name{convert_edgeIndex}
\alias{convert_edgeIndex}
\title{convert_edgeIndex}
\usage{
convert_edgeIndex(edge_list, start, end, db)
}
\arguments{
\item{edge_list}{matrix or data frame columns > 2 ignored}

\item{start}{column name of index matching edge list}

\item{end}{column name of index matching db}

\item{db}{data.frame containing mapping between start and end}
}
\value{
edge list converted from start to end
}
\description{
convert_edgeIndex
}

