% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metabomapr.R
\name{get_KEGG_edgeList}
\alias{get_KEGG_edgeList}
\title{get_KEGG_edgeList}
\usage{
get_KEGG_edgeList(input, el = get_KEGG_pairs())
}
\arguments{
\item{input}{vector of kegg ids}

\item{el}{edge list of KEGG RPAIRs}
}
\description{
get_KEGG_edgeList
}

