% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metabomapr.R
\name{get_KEGG_pairs}
\alias{get_KEGG_pairs}
\title{get_KEGG_pairs}
\usage{
get_KEGG_pairs(type = "main", file = system.file("data/KEGG_RPAIRS", package
  = "metabomapr"))
}
\arguments{
\item{type}{of return 'main' type of reactions or 'full' everything}

\item{file}{RPAIRS list defaults to package data}
}
\description{
get_KEGG_pairs
}

