\name{metabomapr-package}
\alias{metabomapr-package}
\alias{metabomapr}
\docType{package}
\title{
\packageTitle{metabomapr}
}
\description{
\packageDescription{metabomapr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metabomapr}
\packageIndices{metabomapr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{metabomapr}

Maintainer: \packageMaintainer{metabomapr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
