% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metabomapr.R
\name{CID_tanimoto}
\alias{CID_tanimoto}
\title{CID_tanimoto}
\usage{
CID_tanimoto(cids, as = c("adjacency", "edge.list"), ...)
}
\arguments{
\item{cid}{PubChem CID}
}
\description{
CID_tanimoto
}
\details{
... passed to CID_SDF
}

